api-server/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask[async]==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
